% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Save Metadata for a project from TiddlyWiki}
\usage{
get_meta(project, standard_name = TRUE)
}
\arguments{
\item{project}{A character string specifying the project name. Must be a single value.}

\item{standard_name}{A logical indicating whether to assign a standard name based on `aka` values. Default is TRUE.}
}
\value{
A list containing filtered metadata for the project.
}
\description{
Save Metadata for a project from TiddlyWiki
}
\details{
This function retrieves metadata from a TiddlyWiki instance based on a specified project
and applies a filtering process to extract relevant tiddlers. It can optionally map tiddlers
to a standard name based on an alias (`aka` field).
}
