% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{standard_name}
\alias{standard_name}
\title{Get standard names}
\usage{
standard_name(names, group, is_apsim_name = FALSE)
}
\arguments{
\item{names}{A character vector for names}

\item{group}{Group of standard names}

\item{is_apsim_name}{whether to return names only if group is "Variety".}
}
\value{
A vector for standard name if is_apsim_name is FALSE, but a data.frame of
standard_name and apsim_name if is_apsim_name is TRUE.
}
\description{
Get standard names
}
