% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{cache_meta}
\alias{cache_meta}
\title{Cache Metadata}
\usage{
cache_meta(project, as_tibble = TRUE)
}
\arguments{
\item{project}{A character string representing the project name.}

\item{as_tibble}{A logical indicating whether to convert into table. Default is TRUE.}
}
\value{
None. The function saves the metadata to a file.
}
\description{
This function retrieves metadata for a given project
and saves it as an RDS file for caching.
}
