% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{load_meta}
\alias{load_meta}
\title{Load Metadata}
\usage{
load_meta(project, as_tibble = TRUE, update = FALSE)
}
\arguments{
\item{project}{A character string representing the project name.}

\item{as_tibble}{A logical indicating whether to convert into table. Default is TRUE.}

\item{update}{A logical flag; if TRUE, refreshes the metadata cache.}
}
\value{
A list or object containing the project metadata.
}
\description{
This function loads metadata for a given project from cache.
If the metadata file is missing or `update = TRUE`, it refreshes the cache.
}
