% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{standard_name}
\alias{standard_name}
\title{Get standard names}
\usage{
standard_name(names, group)
}
\arguments{
\item{names}{A charactor vector for names}

\item{group}{Group of standard names}
}
\value{
A vector for standard name
}
\description{
Get standard names
}
